requirements.txt
langchain
langchain-ollama
numpy
scipy
pandas
seaborn
matplotlib
scikit-learn

python-dotenv
google-generativeai